Refactor semantic handlers and search infra
Introduce SemanticHandlerMixin to centralize semantic query parsing, validation and fetching (handles VLLM errors and length checks). Move VLLM/Elasticsearch clients into bot/search/infra and update imports accordingly. Consolidate common logic into BotMessageHandler (added _send_top_segment_as_clip, _send_document, _sanitize_filename) and remove duplicated document/sanitize helpers from character/object handlers. Switch character/object lookups to use video_frames_finder, delete legacy finders, and adapt many handlers and responses (including a new get_no_video_path_message and updated semantic frames formatting). Misc: limit search result sizes to settings.MAX_ES_RESULTS and avoid persisting EPISODE-mode semantic searches.
