+++ bot/handlers/character_handler_mixin.py
@@ -9,7 +9,7 @@ from bot.responses.not_sending_videos.characters_handler_responses import (
     get_character_not_found_message,
     parse_character_args,
 )
-from bot.search.character_finder import CharacterFinder
+from bot.search.video_frames_finder import CharacterFinder
 
 
 class CharacterHandlerMixin:
