+++ bot/handlers/sending_videos/character_clip_handler.py
@@ -4,7 +4,6 @@ import math
 from typing import List
 
 from bot.database.database_manager import DatabaseManager
-from bot.database.models import ClipType
 from bot.handlers.bot_message_handler import (
     BotMessageHandler,
     ValidatorFunctions,
@@ -15,13 +14,8 @@ from bot.responses.sending_videos.character_clip_handler_responses import (
     get_log_character_clip_message,
     get_no_quote_provided_message,
     get_no_scenes_found_message,
-    get_no_video_path_message,
 )
-from bot.search.character_finder import CharacterFinder
-from bot.services.scene_snap.scene_snap_service import SceneSnapService
-from bot.settings import settings
-from bot.utils.constants import SegmentKeys
-from bot.video.clips_extractor import ClipsExtractor
+from bot.search.video_frames_finder import CharacterFinder
 
 
 class CharacterClipHandler(CharacterHandlerMixin, BotMessageHandler):
@@ -72,41 +66,8 @@ class CharacterClipHandler(CharacterHandlerMixin, BotMessageHandler):
             segments=json.dumps(segments),
         )
 
-        top_segment = segments[0]
-        if not top_segment.get(SegmentKeys.VIDEO_PATH):
-            await self._reply_error(get_no_video_path_message())
+        if await self._send_top_segment_as_clip(segments[0], series_name):
             return
-
-        start_time = max(0, top_segment[SegmentKeys.START_TIME] - settings.EXTEND_BEFORE)
-        end_time = top_segment[SegmentKeys.END_TIME] + settings.EXTEND_AFTER
-
-        start_time, end_time = await SceneSnapService.snap_clip_times(
-            series_name, top_segment, start_time, end_time, self._logger,
-        )
-
-        clip_duration = end_time - start_time
-        if await self._handle_clip_duration_limit_exceeded(clip_duration):
-            return
-
-        output_filename = await ClipsExtractor.extract_clip(
-            top_segment[SegmentKeys.VIDEO_PATH], start_time, end_time, self._logger,
-        )
-
-        await self._responder.send_video(
-            output_filename,
-            duration=clip_duration,
-            suggestions=["Uzyj /w N aby wybrac inny wynik"],
-        )
-
-        await DatabaseManager.insert_last_clip(
-            chat_id=self._message.get_chat_id(),
-            segment=top_segment,
-            compiled_clip=None,
-            clip_type=ClipType.SINGLE,
-            adjusted_start_time=start_time,
-            adjusted_end_time=end_time,
-            is_adjusted=False,
-        )
 
         await self._log_system_message(
             logging.INFO,
